move keybindings to shared
