+++ src/eval-frame/index.jsx
@@ -20,7 +20,7 @@ import messagePasserEval from "../shared/utils/redux-to-port-message-passer";
 import { iodide } from "./iodide-api/api";
 
 import "./port-to-editor";
-import { initializeDefaultKeybindings } from "../editor/initialization/keybindings";
+import { initializeDefaultKeybindings } from "../shared/keybindings";
 
 messagePasserEval.connectDispatch(store.dispatch);
 window.iodide = iodide;
